refactor update imports, split components
